+++ utils/src/main/java/org/hibernate/tool/util/ReflectHelper.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.internal.util;
+package org.hibernate.tool.util;
 
 public class ReflectHelper {
